Major optimization do codigo dos controllers para melhor compreensao

+++ frontend/controllers/ClientController.php
@@ -9,14 +9,8 @@ use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 
-/**
- * ClientController implements the CRUD actions for Client model.
- */
 class ClientController extends Controller
 {
-    /**
-     * @inheritDoc
-     */
     public function behaviors()
     {
         return array_merge(
@@ -32,64 +26,24 @@ class ClientController extends Controller
         );
     }
 
-    /**
-     * Lists all Client models.
-     *
-     * @return string
-     */
     public function actionIndex()
     {
+        if (!\Yii::$app->user->can('readClient')) {
+            return;
+        }
 
-        $client = Client::findOne([Yii::$app->user->identity->getId()]);
+        $client = Client::findOne([\Yii::$app->user->getId()]);
         return $this->render('index', [
             'client' => $client,
         ]);
     }
 
-    /**
-     * Displays a single Client model.
-     * @param int $user_id User ID
-     * @return string
-     * @throws NotFoundHttpException if the model cannot be found
-     */
-    public function actionView($user_id)
-    {
-        return $this->render('view', [
-            'model' => $this->findModel($user_id),
-        ]);
-    }
-
-    /**
-     * Creates a new Client model.
-     * If creation is successful, the browser will be redirected to the 'view' page.
-     * @return string|\yii\web\Response
-     */
-    public function actionCreate()
-    {
-        $model = new Client();
-
-        if ($this->request->isPost) {
-            if ($model->load($this->request->post()) && $model->save()) {
-                return $this->redirect(['view', 'user_id' => $model->user_id]);
-            }
-        } else {
-            $model->loadDefaultValues();
-        }
-
-        return $this->render('create', [
-            'model' => $model,
-        ]);
-    }
-
-    /**
-     * Updates an existing Client model.
-     * If update is successful, the browser will be redirected to the 'view' page.
-     * @param int $user_id User ID
-     * @return string|\yii\web\Response
-     * @throws NotFoundHttpException if the model cannot be found
-     */
     public function actionUpdate($user_id)
     {
+        if (!\Yii::$app->user->can('updateClient')) {
+            return;
+        }
+
         $model = $this->findModel($user_id);
 
         if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
@@ -101,27 +55,17 @@ class ClientController extends Controller
         ]);
     }
 
-    /**
-     * Deletes an existing Client model.
-     * If deletion is successful, the browser will be redirected to the 'index' page.
-     * @param int $user_id User ID
-     * @return \yii\web\Response
-     * @throws NotFoundHttpException if the model cannot be found
-     */
     public function actionDelete($user_id)
     {
+        if (!\Yii::$app->user->can('updateClient')) {
+            return;
+        }
+
         $this->findModel($user_id)->delete();
 
         return $this->redirect(['index']);
     }
 
-    /**
-     * Finds the Client model based on its primary key value.
-     * If the model is not found, a 404 HTTP exception will be thrown.
-     * @param int $user_id User ID
-     * @return Client the loaded model
-     * @throws NotFoundHttpException if the model cannot be found
-     */
     protected function findModel($user_id)
     {
         if (($model = Client::findOne(['user_id' => $user_id])) !== null) {
@@ -131,3 +75,4 @@ class ClientController extends Controller
         throw new NotFoundHttpException('The requested page does not exist.');
     }
 }
+
